+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This is the changelog file of My Games (games.anikethchavare.com).
 
 **Note:** All the dates below are in the IST time format.
 
+### <i>Commit 2 (6th April 2025)</i>
+
+1. Added a new game "Rock Paper Scissors", in short as "RPS", to the route "/rps".
+
 ### <i>Commit 1 (5th April 2025)</i>
 
 1. Uploaded and deployed "My Games (games.anikethchavare.com)" to GitHub and Vercel for the first time.